// src/app/interfaces/interfaces.ts
export interface RespuestaPais {
  name: name;
  tld?: string[];
  cca2: string;
  ccn3?: string;
  cca3: string;
  cioc?: string;
  independent?: boolean;
  status: string;
  unMember: boolean;
  currencies?: Currencies;
  idd: Idd;
  capital?: string[];
  altSpellings: string[];
  region: string;
  subregion?: string;
  languages?: Languages;
  translations: Translations;
  latlng: number[];
  landlocked: boolean;
  borders?: string[];
  area: number;
  demonyms?: Demonyms;
  flag: string;
  maps: Maps;
  population: number;
  gini?: Gini;
  fifa?: string;
  car: Car;
  timezones: string[];
  continents: string[];
  flags: Flags;
  coatOfArms: CoatOfArms;
  startOfWeek: string;
  capitalInfo: CapitalInfo;
  postalCode?: PostalCode;
}

interface PostalCode {
  format: string;
  regex?: string;
}

interface CapitalInfo {
  latlng?: number[];
}

interface CoatOfArms {
  png?: string;
  svg?: string;
}

interface Flags {
  png: string;
  svg: string;
  alt?: string;
}

interface Car {
  signs?: string[];
  side: string;
}

interface Gini {
  '2010'?: number;
  '2017'?: number;
  '2019'?: number;
  '2009'?: number;
  '2016'?: number;
  '2018'?: number;
  '2015'?: number;
  '2012'?: number;
  '2014'?: number;
  '1992'?: number;
  '1999'?: number;
  '2008'?: number;
  '2003'?: number;
  '2006'?: number;
  '2011'?: number;
  '2004'?: number;
  '2013'?: number;
  '1998'?: number;
  '2005'?: number;
}

interface Maps {
  googleMaps: string;
  openStreetMaps: string;
}

interface Demonyms {
  eng: Eng;
  fra?: Eng;
}

interface Eng {
  f: string;
  m: string;
}

interface Translations {
  ara: Ara;
  bre: Ara;
  ces: Ara;
  cym: Ara;
  deu: Ara;
  est: Ara;
  fin: Ara;
  fra: Ara;
  hrv?: Ara;
  hun: Ara;
  ita: Ara;
  jpn?: Ara;
  kor: Ara;
  nld: Ara;
  per?: Ara;
  pol: Ara;
  por: Ara;
  rus: Ara;
  slk: Ara;
  spa: Ara;
  srp: Ara;
  swe: Ara;
  tur: Ara;
  urd: Ara;
  zho?: Ara;
}

interface Languages {
  ara?: string;
  cal?: string;
  cha?: string;
  eng?: string;
  srp?: string;
  cat?: string;
  aym?: string;
  grn?: string;
  que?: string;
  spa?: string;
  fra?: string;
  hye?: string;
  mfe?: string;
  div?: string;
  smo?: string;
  mkd?: string;
  amh?: string;
  kal?: string;
  arc?: string;
  ckb?: string;
  nld?: string;
  swe?: string;
  est?: string;
  bwg?: string;
  kck?: string;
  khi?: string;
  ndc?: string;
  nde?: string;
  nya?: string;
  sna?: string;
  sot?: string;
  toi?: string;
  tsn?: string;
  tso?: string;
  ven?: string;
  xho?: string;
  zib?: string;
  slk?: string;
  prs?: string;
  pus?: string;
  tuk?: string;
  sag?: string;
  ita?: string;
  lat?: string;
  bis?: string;
  gil?: string;
  zho?: string;
  por?: string;
  tet?: string;
  ssw?: string;
  rus?: string;
  uzb?: string;
  ind?: string;
  hrv?: string;
  pol?: string;
  bos?: string;
  heb?: string;
  pap?: string;
  msa?: string;
  mlt?: string;
  bel?: string;
  tur?: string;
  fin?: string;
  dzo?: string;
  ces?: string;
  cnr?: string;
  nor?: string;
  hin?: string;
  tam?: string;
  mri?: string;
  nzs?: string;
  jam?: string;
  kon?: string;
  lin?: string;
  urd?: string;
  kaz?: string;
  fij?: string;
  hif?: string;
  isl?: string;
  mya?: string;
  ben?: string;
  fil?: string;
  gle?: string;
  nep?: string;
  kor?: string;
  dan?: string;
  tir?: string;
  fas?: string;
  swa?: string;
  ell?: string;
  nfr?: string;
  kin?: string;
  tvl?: string;
  crs?: string;
  khm?: string;
  ukr?: string;
  nau?: string;
  afr?: string;
  deu?: string;
  her?: string;
  hgm?: string;
  kwn?: string;
  loz?: string;
  ndo?: string;
  ton?: string;
  mah?: string;
  de?: string;
  hun?: string;
  bjz?: string;
  ltz?: string;
  fao?: string;
  lua?: string;
  nbl?: string;
  nso?: string;
  zul?: string;
  sin?: string;
  lav?: string;
  ron?: string;
  mon?: string;
  nno?: string;
  nob?: string;
  smi?: string;
  tha?: string;
  nrf?: string;
  lao?: string;
  aze?: string;
  ber?: string;
  bul?: string;
  run?: string;
  sqi?: string;
  kat?: string;
  vie?: string;
  mey?: string;
  jpn?: string;
  lit?: string;
  hat?: string;
  hmo?: string;
  tpi?: string;
  kir?: string;
  pau?: string;
  mlg?: string;
  tkl?: string;
  slv?: string;
  som?: string;
  rar?: string;
  zdj?: string;
  niu?: string;
  gsw?: string;
  roh?: string;
  pih?: string;
  glv?: string;
  tgk?: string;
  pov?: string;
}

interface Idd {
  root?: string;
  suffixes?: string[];
}

interface Currencies {
  JOD?: JOD;
  USD?: JOD;
  RSD?: JOD;
  EUR?: JOD;
  BMD?: JOD;
  BOB?: JOD;
  LYD?: JOD;
  XOF?: JOD;
  AMD?: JOD;
  MUR?: JOD;
  MVR?: JOD;
  MKD?: JOD;
  ETB?: JOD;
  DKK?: JOD;
  IQD?: JOD;
  GTQ?: JOD;
  TTD?: JOD;
  PEN?: JOD;
  SRD?: JOD;
  SEK?: JOD;
  XPF?: JOD;
  FKP?: JOD;
  ZWL?: JOD;
  SAR?: JOD;
  AED?: JOD;
  AFN?: JOD;
  XAF?: JOD;
  PAB?: JOD;
  SYP?: JOD;
  VUV?: JOD;
  HNL?: JOD;
  AUD?: JOD;
  KID?: JOD;
  CLP?: JOD;
  XCD?: JOD;
  MXN?: JOD;
  CNY?: JOD;
  SSP?: JOD;
  SZL?: JOD;
  ZAR?: JOD;
  UZS?: JOD;
  IDR?: JOD;
  PYG?: JOD;
  PLN?: JOD;
  BAM?: BAM;
  CAD?: JOD;
  BRL?: JOD;
  MRU?: JOD;
  ILS?: JOD;
  ANG?: JOD;
  BND?: JOD;
  SGD?: JOD;
  AOA?: JOD;
  BYN?: JOD;
  TRY?: JOD;
  GIP?: JOD;
  BTN?: JOD;
  INR?: JOD;
  VES?: JOD;
  QAR?: JOD;
  CZK?: JOD;
  KWD?: JOD;
  NZD?: JOD;
  JMD?: JOD;
  KYD?: JOD;
  PKR?: JOD;
  KZT?: JOD;
  BHD?: JOD;
  FJD?: JOD;
  ISK?: JOD;
  MMK?: JOD;
  BDT?: JOD;
  PHP?: JOD;
  NPR?: JOD;
  YER?: JOD;
  KRW?: JOD;
  OMR?: JOD;
  ERN?: JOD;
  IRR?: JOD;
  TZS?: JOD;
  SBD?: JOD;
  KES?: JOD;
  DOP?: JOD;
  GBP?: JOD;
  GGP?: JOD;
  RWF?: JOD;
  TVD?: JOD;
  TWD?: JOD;
  GYD?: JOD;
  SCR?: JOD;
  KPW?: JOD;
  BWP?: JOD;
  KHR?: JOD;
  BBD?: JOD;
  COP?: JOD;
  UAH?: JOD;
  NAD?: JOD;
  TOP?: JOD;
  ARS?: JOD;
  CRC?: JOD;
  GHS?: JOD;
  SHP?: JOD;
  EGP?: JOD;
  CUC?: JOD;
  CUP?: JOD;
  HUF?: JOD;
  BZD?: JOD;
  BSD?: JOD;
  AWG?: JOD;
  STN?: JOD;
  RUB?: JOD;
  FOK?: JOD;
  NIO?: JOD;
  TND?: JOD;
  CDF?: JOD;
  NOK?: JOD;
  SLL?: JOD;
  LSL?: JOD;
  DJF?: JOD;
  LKR?: JOD;
  WST?: JOD;
  TMT?: JOD;
  MDL?: JOD;
  CHF?: JOD;
  MWK?: JOD;
  LBP?: JOD;
  MNT?: JOD;
  THB?: JOD;
  NGN?: JOD;
  CVE?: JOD;
  JEP?: JOD;
  DZD?: JOD;
  LAK?: JOD;
  AZN?: JOD;
  MAD?: JOD;
  BGN?: JOD;
  BIF?: JOD;
  UGX?: JOD;
  MZN?: JOD;
  GEL?: JOD;
  RON?: JOD;
  VND?: JOD;
  JPY?: JOD;
  LRD?: JOD;
  HTG?: JOD;
  PGK?: JOD;
  HKD?: JOD;
  KGS?: JOD;
  MGA?: JOD;
  GMD?: JOD;
  MYR?: JOD;
  SOS?: JOD;
  MOP?: JOD;
  ALL?: JOD;
  ZMW?: JOD;
  CKD?: JOD;
  GNF?: JOD;
  KMF?: JOD;
  IMP?: JOD;
  SDG?: BAM;
  TJS?: JOD;
  UYU?: JOD;
}

interface BAM {
  name: string;
}

interface JOD {
  name: string;
  symbol: string;
}

 interface name {
  common: string;
  official: string;
  nativeName?: NativeName;
}

interface NativeName {
  ara?: Ara;
  cal?: Ara;
  cha?: Ara;
  eng?: Ara;
  srp?: Ara;
  cat?: Ara;
  aym?: Ara;
  grn?: Ara;
  que?: Ara;
  spa?: Ara;
  fra?: Ara;
  hye?: Ara;
  mfe?: Ara;
  div?: Ara;
  smo?: Ara;
  mkd?: Ara;
  amh?: Ara;
  kal?: Ara;
  arc?: Ara;
  ckb?: Ara;
  nld?: Ara;
  swe?: Ara;
  est?: Ara;
  bwg?: Ara;
  kck?: Ara;
  khi?: Ara;
  ndc?: Ara;
  nde?: Ara;
  nya?: Ara;
  sna?: Ara;
  sot?: Ara;
  toi?: Ara;
  tsn?: Ara;
  tso?: Ara;
  ven?: Ara;
  xho?: Ara;
  zib?: Ara;
  slk?: Ara;
  prs?: Ara;
  pus?: Ara;
  tuk?: Ara;
  sag?: Ara;
  ita?: Ara;
  lat?: Ara;
  bis?: Ara;
  gil?: Ara;
  zho?: Ara;
  por?: Ara;
  tet?: Ara;
  ssw?: Ara;
  rus?: Ara;
  uzb?: Ara;
  ind?: Ara;
  hrv?: Ara;
  pol?: Ara;
  bos?: Ara;
  heb?: Ara;
  pap?: Ara;
  msa?: Ara;
  mlt?: Ara;
  bel?: Ara;
  tur?: Ara;
  fin?: Ara;
  dzo?: Ara;
  ces?: Ara;
  cnr?: Ara;
  nor?: Ara;
  hin?: Ara;
  tam?: Ara;
  mri?: Ara;
  nzs?: Ara;
  jam?: Ara;
  kon?: Ara;
  lin?: Ara;
  urd?: Ara;
  kaz?: Ara;
  fij?: Ara;
  hif?: Ara;
  isl?: Ara;
  mya?: Ara;
  ben?: Ara;
  fil?: Ara;
  gle?: Ara;
  nep?: Ara;
  kor?: Ara;
  dan?: Ara;
  tir?: Ara;
  fas?: Ara;
  swa?: Ara;
  ell?: Ara;
  nfr?: Ara;
  kin?: Ara;
  tvl?: Ara;
  crs?: Ara;
  khm?: Ara;
  ukr?: Ara;
  nau?: Ara;
  afr?: Ara;
  deu?: Ara;
  her?: Ara;
  hgm?: Ara;
  kwn?: Ara;
  loz?: Ara;
  ndo?: Ara;
  ton?: Ara;
  mah?: Ara;
  bar?: Ara;
  hun?: Ara;
  bjz?: Ara;
  ltz?: Ara;
  fao?: Ara;
  lua?: Ara;
  nbl?: Ara;
  nso?: Ara;
  zul?: Ara;
  sin?: Ara;
  lav?: Ara;
  ron?: Ara;
  mon?: Ara;
  nno?: Ara;
  nob?: Ara;
  smi?: Ara;
  tha?: Ara;
  nrf?: Ara;
  lao?: Ara;
  aze?: Ara;
  ber?: Ara;
  bul?: Ara;
  run?: Ara;
  sqi?: Ara;
  kat?: Ara;
  vie?: Ara;
  mey?: Ara;
  jpn?: Ara;
  lit?: Ara;
  hat?: Ara;
  hmo?: Ara;
  tpi?: Ara;
  kir?: Ara;
  pau?: Ara;
  mlg?: Ara;
  tkl?: Ara;
  slv?: Ara;
  som?: Ara;
  rar?: Ara;
  zdj?: Ara;
  niu?: Ara;
  gsw?: Ara;
  roh?: Ara;
  pih?: Ara;
  glv?: Ara;
  tgk?: Ara;
  pov?: Ara;
}

interface Ara {
  official: string;
  common: string;
}